Refactoring out a settings.js file.
